Commit for #31 & #54
- Tidied up Build Assistant Project
- Added READ-ME to Launcher
