+++ LPG Launcher/VersionAutoIncrementer.cs
@@ -1,6 +1,6 @@
 ﻿
 using System.Reflection;
 
-[assembly: AssemblyVersion("0.1.3.664")]
-[assembly: AssemblyFileVersion("0.1.3.664")]
+[assembly: AssemblyVersion("0.1.3.896")]
+[assembly: AssemblyFileVersion("0.1.3.896")]
